Move lib integration tests
